feat(jdbc): 新增支持mysql/oracle配置

+++ spring5x-jdbc-database/src/main/java/com/zja/rowmapper/UserEntityRowMapper.java
@@ -13,7 +13,7 @@ public class UserEntityRowMapper implements RowMapper<UserEntity> {
     @Override
     public UserEntity mapRow(ResultSet resultSet, int i) throws SQLException {
         UserEntity userEntity = new UserEntity();
-        userEntity.setId(resultSet.getInt("id"));
+        userEntity.setId(resultSet.getLong("id"));
         userEntity.setUserName(resultSet.getString("username"));
         userEntity.setAge(resultSet.getInt("age"));
         return userEntity;
